Substitute calls to `apply()`
In the function that uses arrays as input, the calls to apply() to
compute summary statistics are relatively slow.  This is a bit faster
than the previous version, although it has to create lists of matrices
by month for the calls to pmin() and pmax(), which adds a little bit
of overhead.